make gtk3 performance usable
only render into the backbuffer once, don't re-render on "draw".

In order that make that possible the back-buffer has to be refilled
on resize events and a whole pile of things had to be fixed in the
underlying svp backend.

now it scrolls like the wind, well comparitively anyway, and while
you probably wouldn't want to use it yet, you could in a pinch

Change-Id: I8938fa099e5594fd44c32baaad9043946280212c