android: make & build appcompat-v7 library, up target to v21
Previously we included the android support library v4 for some
GUI elements like GridView. This commit in addition adds the v7
appcompat library which is needed for the new Lollipop style
Toolbar and many other new GUI elements.

The appcompat v7 library is not distributed as only a jar file
but needs to be build (as it includes additional resources) and
included as a library project. So to do this the content is copied
from SDK and build with the build system. The files also include
the v4 so it doesn't need to be copied from SDK anymore.

The target had to be raised to v21 (Lollipop), however the minimum
SDK version remains unchanged.

Change-Id: I4f1a6ce69e7f6c3f9df784a6835f376a01d4dfdb