Fix signature overflow check in the NSS case
We didn't actually check this correctly at all, but gladly overwrote the
allocated part of the output PDF, thus obviously rendering it invalid.

The parameter passed to PORT_NewArea is a default chunk size, not a maximum
anything, so it was misleading, even if not wrong as such, to pass
MAX_SIGNATURE_CONTENT_LENGTH to it. Use 10000 instead.

No need to do the overflow check twice in the Win32 case.

Change-Id: Ifa796dbb74b32e857f7184c1e8ada97ba124b020