Speed up ScriptRun::getPairIndex
Noted as a hot spot in the callgrind of
test-files/loperf/091_fdo#39372.odt

The idea for the lookup tables is credited to mjayfrancis.

This takes it from 1.8 billion instructions to 1.1 billion instructions.

This change has been submitted upstream (along with other changes
to these source files that have not yet been pushed there)
in ICU ticket:
    http://bugs.icu-project.org/trac/ticket/11577

Change-Id: I917bdea4263320f9e88207657b7f3140d95ae2b8
Reviewed-on: https://gerrit.libreoffice.org/14834
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>