Follow-up to 01ebb817ce264f1aa286d5c838f2ca78c0712436
To avoid losing functionality (components missing at run-time), we need to add
the corrsponding new-style functions instead of the old-style one that was
removed. In this case it was just one.

(In those cases where one old-style is replaces with several new-style ones,
ideally one can then avoid some of the replacements in the iOS and Android
case, and save on code size.)

Change-Id: I6d4a90e84600debdf62b47408264c439a6f0e312