tdf#90594: MM: Fix for broken translation between Bookmarks and row indizes
The datasets used in the unomailmerge service can be selected via a
property "Selection". This property expects a sequence of sdbc::Bookmark
elements that need to be translated into row-inizes. This translation
was broken - the loop-pointer pTranslated was never incremented which
caused a wrong "result" selection in which only the first element was
changed.

Change-Id: I319131caa89e327b11406ab7d0c56a833385d433
Reviewed-on: https://gerrit.libreoffice.org/15289
Reviewed-by: David Tardon <dtardon@redhat.com>
Tested-by: David Tardon <dtardon@redhat.com>