Reverts and correct "tdf#88230: Cleanup solar mutex yielding"
This reverts commit b26ab2115462e65a9f5f478151bf8118294b0d67. This
commit directly used SolarMutexReleaser class which releases the solar
mutex first and then acquires the mutex after the destruction of the
object. This is not the correct way of replacing the pre-existing code
where mutex needs to be acquired first and then released.

Additionally, this commit adds a correct replacement for pre-existing
code too.

Change-Id: I79aff608d94ed2d4736679a691b8c18f8a23c581
Reviewed-on: https://gerrit.libreoffice.org/14981
Reviewed-by: Noel Grandin <noelgrandin@gmail.com>
Tested-by: Noel Grandin <noelgrandin@gmail.com>