lokdocview: sync graphic move behavior with Android
Android behavior is: if there is a graphic selection, then drag inside
the shape is a move.

lokdocivew behavior was: drag on the border of a graphic selection is a
move, but inside, it's just a pair of clicks.

Since commit 7a6db6a34e30235da8a5e0f31b6c50b75e14e68a (SdrMarkView tiled
rendering: suppress handles during text edit, 2015-03-19), we can depend
on the availability of the handles to decide if we want to move the
shape or do text editing, so do the same in lokdocview as well.

Change-Id: I2b105089e9acbbda75568622d4f2b1adecfe08f1