why would we ever want FLOATWIN_POPUPMODE_PATHMOUSECANCELCLICK
if we're from the menubar we never set it, but if we are
a submenu from a menu in the menubar we set it.

with it set, and on right click, the right click is passed
through to what ever is underneath the pointer

I don't see why we want to do that, especially the inconsistency
from e.g. the File menu, vs the "File->New" menu

Change-Id: I548f5c95e2511b5e921e8b74600f65eecdd8f498