tdf#90298 rescan for bundled extensions when installing Mac langpack
LO only scans for new extensions when the folder's timestamp is newer
than a timestamp file in the user's profile. When installing a
languagepack, the folder is set to the build-time of the package. touch
the dir to set it to install time and thus allow LO to pick up newly
installed dictionaries and thesaurus/hyphenation data.

Change-Id: I888f830d2325774cd98e3624c19d2be19d4d6db2