SwViewShell::PaintTile: avoid incomplete tiles when layout is not ready
The problem was that in SwViewShell::Paint(), in case
CheckInvalidForPaint() returns true, we don't paint anything, but rely
on invalidations to call Paint() again and hand out an incomplete tile.

For tiled rendering, we want to hand out complete tiles all the time, so
call CheckInvalidForPaint() early in SwViewShell::PaintTile() that will
ensure that by the time SwViewShell::Paint() is called, the layout is
ready for that tile.

Without this, when opening any larger document with the Android viewer,
about at least once out of 5 times an incomplete tile (e.g. almost
totally white) was shown for the first page, and the user had to work
around this by zooming out + in again to force a re-draw with valid
layout.

Change-Id: Ic7d3a6ddbf9f04beae9ecaa9563d3f5b4d22cc5c