MM: fixed invalid UNO-Reference as return value of UNO MailMergeService
This fix is regarding cc34b1891 "MM: export the SwDocShell mail merge via UNO":

In case of using the new output type MailMergeType.SHELL, the UNO
MailMergeService returned a XTextDocument instance that was not
correctly initialized. In particular after calling doc.getCurrentController()
(with doc = the returned instance) we recieved NULL. The reason was, that the
previous implementation used to create a new SwXTextDocument instance
instead of just returning the BaseModel that is already associated with
the internal SwDocShell. This is fixed in this patch.

I also tested this patch with the unit test sw/CppunitTest_sw_mailmerge
and everything seems to be fine now. doc.getCurrentController() no longer
returns NULL.

Change-Id: Ic116f8f25ab9686ef56950e97ba202a55ab81fb8
Reviewed-on: https://gerrit.libreoffice.org/15207
Tested-by: Michael Stahl <mstahl@redhat.com>
Reviewed-by: Michael Stahl <mstahl@redhat.com>