SwEditWin tiled rendering: do word selection even if the cursor was corrected
When tiled rendering, we always want to create some kind of selection
when long pushing on a table cell. If there is at least one word in the
cell, then we want to select the nearest cell, and when it's empty, then
we want to select the cell itself.

In case there is a table with 1 rows and 2 columns, and A1 is empty, but
A2 has multiple lines, then in A1 there is an area that's outside the
text frame of the empty paragraph, but inside the A1 cell frame. When
clicking on that area, the cursor position gets corrected. With this
change, we get a proper selection on long push even when pushing on that
"outside text, inside cell" area, too.

Change-Id: Ic61743014708f127087243cb7b129f8abd72edaa