initial load of bibliography doesn't set bibliography type
moving to a new row will set the new type, but initially
its not set because the ListStore property was set after
the model is connected to the database so at the initial
connection time there isn't a available set of values
to map the row to.

So move that connection code from general.cxx to datman.cxx
and put it before the connection is established

we can then remove the intermediate aBibTypeArr list and
those resources then need to be global bib ones.

Change-Id: I12276c3df9955ec399e1b5de6d0836c69ebb1ebb