tdf#71278 Revert both the previous cursor commit and the commit it fixed
The original commit was based on a misunderstanding. In fact, the call to
GetSalData()->getCursor() at the end of AquaSalFrame::getCurrentCursor()
already handles POINTER_NULL correctly, or at least in the same way as on
other platforms. The cursor in presentation mode disappears after 10
seconds - if this is to be changed, it should be changed for all platforms.

Reverts:

  commit cf60ac74503231227a99e01d4555dd198747ea7c
  Author: Matthew Francis <mjay.francis@gmail.com>
  Date:   Sun Apr 12 23:10:58 2015 +0800

      tdf#71278 Set an invisible cursor rather than hiding the cursor

  commit ca2d92f515a04c6fba616e4f0ec9bd753508b316
  Author: jorendc <joren.libreoffice@telenet.be>
  Date:   Sun Apr 14 00:33:19 2013 +0200

      fdo#31082 cursor visible during impress presentation

Change-Id: Ie40882661dc1a2f4857027ff04393874042e04ec
Reviewed-on: https://gerrit.libreoffice.org/15316
Tested-by: Norbert Thiebaud <nthiebaud@gmail.com>
Reviewed-by: Norbert Thiebaud <nthiebaud@gmail.com>