The mode lines says we use 4-column indentation steps, so do that then
Also, be consistent in spacing, indentation and alignment. Inconsistency is
infuritating.

In this case, I decided to drop the attempts to align the names of "methods",
and their parameter lists. They were not all aligned anyway, and some of those
that tried to be in fact were off by one column. Also, consistently use a
blank line between the method pointers,

If you feel strongly about this and *want* such inter-line alignment, by all
means re-introduce it, but at least try to be consistent then, and do it for
all methods.

Change-Id: I382ca0eaf8533ebc3f829ebd978a9676ebcefaf0