RTF import: fix handling of \objdata
There were two problems here:

1) The input stream is read multiple times, so plain naive XInputStream
is not enough, XSeekable is needed, as it is the case for the stream
provided by the OOXML tokenizer.

2) Seeking to the correct start position is not enough, code using the
stream will assume that seeking to the beginning of the stream allows
reading the correct data again, so provide a dedicated stream instead.

With this, handling of math equations created by Word <= 2007 and
embedded to RTF files can be edited finally.

Change-Id: Ic225e1e1060f8bdd5651a21e68970620c9ac6b68