unify code duplication in ScInterpreter::ScSum and IterateParameters
The code in ScSum is mostly identical with that in IterateParameters.
Worse, functions like SUBTOTAL and AGGREGATE use IterateParameters for
the SUM function.
Now we have one set of code for all SUM-functions and averted the danger of
diverting pieces of code which are supposed to do the same thing.

Removed old-style bNull-method too.

Change-Id: Ia4f22c21c4b3c3a244ea1cd10c30d8cfcaa8ef6c
Reviewed-on: https://gerrit.libreoffice.org/15362
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Eike Rathke <erack@redhat.com>