Resolves: tdf#90681 table model can have truly empty cells
old school complex tables can create a table model
where a cell exists in the table but there are no
paragraphs in it.

--------
| A | C |
----
| B |*D*|
--------

i.e. normally for the above there are 4 nodes which
get exported, even though C and D are merged. But it
can be the case that there are only three nodes, and
*D* is missing

be conservative for now and only do this for the
obviously broken no cell start but cell end case
and incrementally build up the test-cases

Change-Id: I5703595f61688a66b7fac7f3905ace0c207c9875