fdo#71797 strange font selection bug with font preview
The dialog seems to be automatically in preview mode from the start
(which seems wrong) That bogus mode of operation (which I failed to fix
from the begining) is somewhat responsible for the behaviour we see.
Adding here a further ugly hack to ensure we use GetText when not in
drop down mode (this is when clicking on a cell or cycling through the
font list with up/down keys) In dropdown mode sweeping with the mouse
highlights fonts (and we preview the highlighted ones) Note: in this
case the selected font and the highlighted font are different, we
need to pick up the correct one.

Change-Id: Id2ba96a57d2f8cfc9657bce8de3ac27588868eaf
Signed-off-by: Noel Power <noel.power@suse.com>