tdf#90720 Textframe background was not transparent
In case of 'no fill' for vml textboxes, MS office rendered as
transpanrent textbox.
When opened same file on LO, it rendered it as textframe with white
background. (Even if the fillstyle is type of 'no fill' and 'backcolortransparency
' is 100%)
So, Fixing this issue by applying 100% 'FillTransparency' if it gets above situation.

Change-Id: Ied2d3855354ada994288c29c78d9ccb4e5b0b4d6
Reviewed-on: https://gerrit.libreoffice.org/15390
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>