tdf#80708 RTF export: fix unexpected table row break
Regression from commit 20171845e582d213f6e530bd55e4fd36c61ae55c
(n#203704 fix RTF export table output for subtables, 2012-02-08), the
problem is that RtfAttributeOutput::StartTableRow() has two callers
depending on if it's the first or a later row, but the m_bTableRowEnded
flag storing if that method is called is only set in one case.

Fix the problem by moving the update of the flag to the method itself.

Change-Id: Ieae1fb53f9528fc74d037d2fd783b3979b17cab8