RTF import: handle d{x,y}WrapDist* picture properties
Commit ec91f2ae290aec39d627590ee9292bc1dca01377 (RTF import of
d{x,y}WrapDist* shape properties, 2013-04-09) added support for wrap
distance of shapes, but that was ignored for shapes, as dmapper later
overwrote the set margins.

Fix this by generating the expected tokens in case of pictures, then
dmapper will take care of the rest.

Also add testcases for the original shape wrap distance feature that was
missing.

Change-Id: I6f219ee6fef71328368409d142897dbae77a0f2f